Start renaming project
With the move to the new repository, we are no longer pretending to be a
fork of the LCM types from libbot (which already we weren't, really, as
we have been adding additional types for some time). Accordingly, start
renaming the project files and build artifacts to reflect our new
identity.