Fix race condition building C library
Update build logic to use the new functions from LCM which prevent a
possible race condition between generating the LCM type bindings and
trying to compile the C source file for a type that refers to other
types.

Also, update the minimum CMake version to 3.1. (Strictly speaking, our
creation of an `INTERFACE` library already required CMake 3.1 or later,
so the old requirement was actually a lie. The new LCM functions we are
using are only defined for CMake 3.1 or later.)